First version

+++ src/main/scala/com/codenjoy/dojo/bomberman/client/package.scala
@@ -4,7 +4,7 @@ import com.codenjoy.dojo.bomberman.model.Elements
 
 package object client {
 
-  type Score = Int
+  type Score = Double
   type Element = Elements
 
 }
